Process the cells from the top left corner
The cells of the grid need to be processed in a specific order to
ensure that the links come out as expected. Change the code to create
the links in order according to the following:

Consider the table as a digraph where each square is a node and there
is an edge going out of each square to the the square to the right and
below it. If viewed as a tree we need to process these nodes breadth-first